+++ python/cugraph/tests/test_betweenness_centrality.py
@@ -37,6 +37,7 @@ with warnings.catch_warnings():
 # Parameters
 # =============================================================================
 DIRECTED_GRAPH_OPTIONS = [False, True]
+WEIGHTED_GRAPH_OPTIONS = [False, True]
 ENDPOINTS_OPTIONS = [False, True]
 NORMALIZED_OPTIONS = [False, True]
 DEFAULT_EPSILON = 0.0001
@@ -64,7 +65,7 @@ def calc_betweenness_centrality(
     result_dtype=np.float64,
     use_k_full=False,
     multi_gpu_batch=False,
-
+    edgevals=False,
 ):
     """ Generate both cugraph and networkx betweenness centrality
 
@@ -103,6 +104,10 @@ def calc_betweenness_centrality(
     multi_gpu_batch : bool
         When True, enable mg batch after constructing the graph
 
+    edgevals: bool
+        When True, enable tests with weighted graph, should be ignored
+        during computation.
+
     Returns
     -------
 
@@ -115,7 +120,9 @@ def calc_betweenness_centrality(
     G = None
     Gnx = None
 
-    G, Gnx = utils.build_cu_and_nx_graphs(graph_file, directed=directed)
+    G, Gnx = utils.build_cu_and_nx_graphs(graph_file, directed=directed,
+                                          edgevals=edgevals)
+
     assert G is not None and Gnx is not None
     if multi_gpu_batch:
         G.enable_batch()
@@ -303,6 +310,7 @@ def prepare_test():
 @pytest.mark.parametrize("endpoints", ENDPOINTS_OPTIONS)
 @pytest.mark.parametrize("subset_seed", SUBSET_SEED_OPTIONS)
 @pytest.mark.parametrize("result_dtype", RESULT_DTYPE_OPTIONS)
+@pytest.mark.parametrize("edgevals", WEIGHTED_GRAPH_OPTIONS)
 def test_betweenness_centrality(
     graph_file,
     directed,
@@ -312,6 +320,7 @@ def test_betweenness_centrality(
     endpoints,
     subset_seed,
     result_dtype,
+    edgevals
 ):
     prepare_test()
     sorted_df = calc_betweenness_centrality(
@@ -323,6 +332,7 @@ def test_betweenness_centrality(
         endpoints=endpoints,
         seed=subset_seed,
         result_dtype=result_dtype,
+        edgevals=edgevals,
     )
     compare_scores(sorted_df, first_key="cu_bc", second_key="ref_bc")
 
@@ -336,6 +346,7 @@ def test_betweenness_centrality(
 @pytest.mark.parametrize("subset_seed", SUBSET_SEED_OPTIONS)
 @pytest.mark.parametrize("result_dtype", RESULT_DTYPE_OPTIONS)
 @pytest.mark.parametrize("use_k_full", [True])
+@pytest.mark.parametrize("edgevals", WEIGHTED_GRAPH_OPTIONS)
 def test_betweenness_centrality_k_full(
     graph_file,
     directed,
@@ -346,6 +357,7 @@ def test_betweenness_centrality_k_full(
     subset_seed,
     result_dtype,
     use_k_full,
+    edgevals
 ):
     """Tests full betweenness centrality by using k = G.number_of_vertices()
     instead of k=None, checks that k scales properly"""
@@ -360,6 +372,7 @@ def test_betweenness_centrality_k_full(
         seed=subset_seed,
         result_dtype=result_dtype,
         use_k_full=use_k_full,
+        edgevals=edgevals
     )
     compare_scores(sorted_df, first_key="cu_bc", second_key="ref_bc")
 
@@ -376,6 +389,7 @@ def test_betweenness_centrality_k_full(
 @pytest.mark.parametrize("endpoints", ENDPOINTS_OPTIONS)
 @pytest.mark.parametrize("subset_seed", [None])
 @pytest.mark.parametrize("result_dtype", RESULT_DTYPE_OPTIONS)
+@pytest.mark.parametrize("edgevals", WEIGHTED_GRAPH_OPTIONS)
 def test_betweenness_centrality_fixed_sample(
     graph_file,
     directed,
@@ -385,6 +399,7 @@ def test_betweenness_centrality_fixed_sample(
     endpoints,
     subset_seed,
     result_dtype,
+    edgevals
 ):
     """Test Betweenness Centrality using a subset
 
@@ -400,6 +415,7 @@ def test_betweenness_centrality_fixed_sample(
         endpoints=endpoints,
         seed=subset_seed,
         result_dtype=result_dtype,
+        edgevals=edgevals
     )
     compare_scores(sorted_df, first_key="cu_bc", second_key="ref_bc")
 
@@ -412,6 +428,7 @@ def test_betweenness_centrality_fixed_sample(
 @pytest.mark.parametrize("endpoints", ENDPOINTS_OPTIONS)
 @pytest.mark.parametrize("subset_seed", SUBSET_SEED_OPTIONS)
 @pytest.mark.parametrize("result_dtype", RESULT_DTYPE_OPTIONS)
+@pytest.mark.parametrize("edgevals", WEIGHTED_GRAPH_OPTIONS)
 def test_betweenness_centrality_weight_except(
     graph_file,
     directed,
@@ -421,6 +438,7 @@ def test_betweenness_centrality_weight_except(
     endpoints,
     subset_seed,
     result_dtype,
+    edgevals
 ):
     """Calls betwenness_centrality with weight
 
@@ -438,6 +456,7 @@ def test_betweenness_centrality_weight_except(
             endpoints=endpoints,
             seed=subset_seed,
             result_dtype=result_dtype,
+            edgevals=edgevals
         )
         compare_scores(sorted_df, first_key="cu_bc", second_key="ref_bc")
 
@@ -450,6 +469,7 @@ def test_betweenness_centrality_weight_except(
 @pytest.mark.parametrize("endpoints", ENDPOINTS_OPTIONS)
 @pytest.mark.parametrize("subset_seed", SUBSET_SEED_OPTIONS)
 @pytest.mark.parametrize("result_dtype", [str])
+@pytest.mark.parametrize("edgevals", WEIGHTED_GRAPH_OPTIONS)
 def test_betweenness_invalid_dtype(
     graph_file,
     directed,
@@ -459,6 +479,7 @@ def test_betweenness_invalid_dtype(
     endpoints,
     subset_seed,
     result_dtype,
+    edgevals
 ):
     """Test calls edge_betwenness_centrality an invalid type"""
 
@@ -473,15 +494,22 @@ def test_betweenness_invalid_dtype(
             endpoints=endpoints,
             seed=subset_seed,
             result_dtype=result_dtype,
+            edgevals=edgevals
         )
         compare_scores(sorted_df, first_key="cu_bc", second_key="ref_bc")
 
 
 @pytest.mark.parametrize("graph_file", utils.DATASETS_SMALL)
-def test_betweenness_centrality_nx(graph_file):
+@pytest.mark.parametrize("directed", DIRECTED_GRAPH_OPTIONS)
+@pytest.mark.parametrize("edgevals", WEIGHTED_GRAPH_OPTIONS)
+def test_betweenness_centrality_nx(
+        graph_file,
+        directed,
+        edgevals
+):
     prepare_test()
 
-    Gnx = utils.generate_nx_graph_from_file(graph_file)
+    Gnx = utils.generate_nx_graph_from_file(graph_file, directed, edgevals)
 
     nx_bc = nx.betweenness_centrality(Gnx)
     cu_bc = cugraph.betweenness_centrality(Gnx)